feat(look&feel,apollo): fix tests for contentItemDuoAction
